An initial Drupal login Locust script
Logs the user in (randomly picked from a predefined list of users) and navigates to either the frontpage or the about page.
It also fetches any static assets to simulate a real browsing experience. As the static assets are sometimes private files
and these cannot always be cached it is useful to also test this.